# public/swagger.json
{
	"components": {
...
		"schemas": {
...
			"InternalServerError": {
				"properties": {
					"msg": {
						"type": "string"
					},
					"ok": {
						"type": "boolean"
					},
					"error": {},
					"status": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"msg",
					"ok",
					"error",
					"status"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"NotFoundItems": {
				"properties": {
					"msg": {
						"type": "string"
					},
					"ok": {
						"type": "boolean"
					},
					"status": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"msg",
					"ok",
					"status"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"Empleado": {
				"properties": {
					"empleado_id": {
						"type": "number",
						"format": "double"
					},
					"persona_id": {
						"type": "number",
						"format": "double"
					},
					"cargo": {
						"type": "string"
					},
					"salario": {
						"type": "number",
						"format": "double"
					},
					"fecha_inicio_contrato": {
						"type": "string"
					},
					"supervisor_id": {
						"type": "number",
						"format": "double",
						"nullable": true
					},
					"fecha_creacion": {
						"type": "string"
					},
					"fecha_actualizacion": {
						"type": "string"
					},
					"empresa_id": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"empleado_id",
					"persona_id",
					"cargo",
					"salario",
					"fecha_inicio_contrato",
					"supervisor_id",
					"fecha_creacion",
					"fecha_actualizacion",
					"empresa_id"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"CreateTareaResponse": {
				"properties": {
					"ok": {
						"type": "boolean"
					},
					"msg": {
						"type": "string"
					},
					"status": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"ok",
					"msg",
					"status"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CreateTareaData": {
				"properties": {
					"empleado_id": {
						"type": "number",
						"format": "double"
					},
					"supervisor_id": {
						"type": "number",
						"format": "double"
					},
					"descripcion": {
						"type": "string"
					},
					"fecha": {
						"type": "string",
						"format": "date-time"
					},
					"prioridad": {
						"type": "string",
						"enum": [
							"alta",
							"media",
							"baja"
						]
					},
					"estado": {
						"type": "string",
						"enum": [
							"en",
							"pe",
							"de",
							"co",
							"su",
							"ca"
						]
					}
				},
				"required": [
					"empleado_id",
					"descripcion",
					"fecha",
					"prioridad",
					"estado"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UpdateTareaResponse": {
				"properties": {
					"ok": {
						"type": "boolean"
					},
					"msg": {
						"type": "string"
					},
					"status": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"ok",
					"msg",
					"status"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UpdateTareaData": {
				"properties": {
					"descripcion": {
						"type": "string"
					},
					"fecha": {
						"type": "string",
						"format": "date-time"
					},
					"prioridad": {
						"type": "string",
						"enum": [
							"alta",
							"media",
							"baja"
						]
					},
					"estado": {
						"type": "string",
						"enum": [
							"en",
							"pe",
							"de",
							"co",
							"su",
							"ca"
						]
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"DeleteTareaResponse": {
				"properties": {
					"ok": {
						"type": "boolean"
					},
					"msg": {
						"type": "string"
					},
					"status": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"ok",
					"msg",
					"status"
				],
				"type": "object",
				"additionalProperties": false
			},
			"TareaWithRelations": {
				"properties": {
					"tarea_id": {
						"type": "number",
						"format": "double"
					},
					"descripcion": {
						"type": "string"
					},
					"fecha": {
						"type": "string",
						"format": "date-time"
					},
					"prioridad": {
						"type": "string",
						"enum": [
							"alta",
							"media",
							"baja"
						]
					},
					"estado": {
						"type": "string",
						"enum": [
							"en",
							"pe",
							"de",
							"co",
							"su",
							"ca"
						]
					},
					"empleado": {
						"$ref": "#/components/schemas/Empleado"
					},
					"supervisor": {
						"$ref": "#/components/schemas/Empleado"
					}
				},
				"required": [
					"tarea_id",
					"descripcion",
					"fecha",
					"prioridad",
					"estado",
					"empleado",
					"supervisor"
				],
				"type": "object",
				"additionalProperties": false
			},
			"GetTareasResponse": {
				"properties": {
					"ok": {
						"type": "boolean"
					},
					"data": {
						"items": {
							"$ref": "#/components/schemas/TareaWithRelations"
						},
						"type": "array"
					},
					"status": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"ok",
					"data",
					"status"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"SuccessResponseCreateLoanRequest": {
				"properties": {
					"ok": {
						"type": "boolean"
					},
					"msg": {
						"type": "string"
					},
					"status": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"ok",
					"msg",
					"status"
				],
				"type": "object",
				"additionalProperties": false
			},
			"LoanRequestDocument": {
				"properties": {
					"nombre": {
						"type": "string"
					},
					"enlace": {
						"type": "string"
					}
				},
				"required": [
					"nombre",
					"enlace"
				],
				"type": "object",
				"additionalProperties": false
			},
			"LoanRequest": {
				"properties": {
					"cliente_id": {
						"type": "number",
						"format": "double"
					},
					"tipo_prestamo_id": {
						"type": "number",
						"format": "double"
					},
					"empresa_id": {
						"type": "number",
						"format": "double"
					},
					"monto_solicitado": {
						"type": "number",
						"format": "double"
					},
					"emp_id": {
						"type": "number",
						"format": "double"
					},
					"documentos": {
						"items": {
							"$ref": "#/components/schemas/LoanRequestDocument"
						},
						"type": "array"
					}
				},
				"required": [
					"cliente_id",
					"tipo_prestamo_id",
					"empresa_id",
					"monto_solicitado",
					"emp_id",
					"documentos"
				],
				"type": "object",
				"additionalProperties": false
			},
			"SuccessResponse": {
				"properties": {
					"ok": {
						"type": "boolean"
					},
					"msg": {
						"type": "string"
					},
					"status": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"ok",
					"msg",
					"status"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Partial_LoanRequest_": {
				"properties": {
					"cliente_id": {
						"type": "number",
						"format": "double"
					},
					"tipo_prestamo_id": {
						"type": "number",
						"format": "double"
					},
					"empresa_id": {
						"type": "number",
						"format": "double"
					},
					"monto_solicitado": {
						"type": "number",
						"format": "double"
					},
					"emp_id": {
						"type": "number",
						"format": "double"
					},
					"documentos": {
						"items": {
							"$ref": "#/components/schemas/LoanRequestDocument"
						},
						"type": "array"
					}
				},
				"type": "object",
				"description": "Make all properties in T optional"
			},
			"CampoPlantillaCSV": {
				"properties": {
					"campo_id": {
						"type": "number",
						"format": "double"
					},
					"plantilla_id": {
						"type": "number",
						"format": "double"
					},
					"titulo_campo": {
						"type": "string"
					},
					"identificador_campo": {
						"type": "string"
					}
				},
				"required": [
					"campo_id",
					"plantilla_id",
					"titulo_campo",
					"identificador_campo"
				],
				"type": "object",
				"additionalProperties": false
			},
			"PlantillaCSV": {
				"properties": {
					"plantilla_id": {
						"type": "number",
						"format": "double"
					},
					"banco_id": {
						"type": "number",
						"format": "double"
					},
					"empresa_id": {
						"type": "number",
						"format": "double"
					},
					"titulo_plantilla": {
						"type": "string"
					},
					"campos": {
						"items": {
							"$ref": "#/components/schemas/CampoPlantillaCSV"
						},
						"type": "array"
					}
				},
				"required": [
					"plantilla_id",
					"banco_id",
					"empresa_id",
					"titulo_plantilla",
					"campos"
				],
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "finantek_api",
		"version": "0.0.1",
		"description": "API fir finantek",
		"license": {
			"name": "ISC"
		},
		"contact": {
			"name": "Daniel Valdez | hello@daniel-valdez.com | https://daniel-valdez.com"
		}
	},
	"openapi": "3.0.0",
	"paths": {
...
		"/api/tareas/crear": {
			"post": {
				"operationId": "CrearTarea",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/CreateTareaResponse"
										},
										{
											"$ref": "#/components/schemas/InternalServerError"
										}
									]
								}
							}
						}
					},
					"500": {
						"description": "Internal Server Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/InternalServerError"
								},
								"examples": {
									"Example 1": {
										"value": {
											"ok": false,
											"msg": "Error interno del sistema, por favor contacte al administrador del sistema",
											"error": {},
											"status": 500
										}
									}
								}
							}
						}
					}
				},
				"tags": [
					"Tareas"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateTareaData"
							}
						}
					}
				}
			}
		},
		"/api/tareas/update/{tarea_id}": {
			"put": {
				"operationId": "ActualizarTarea",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/UpdateTareaResponse"
										},
										{
											"$ref": "#/components/schemas/InternalServerError"
										}
									]
								}
							}
						}
					},
					"500": {
						"description": "Internal Server Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/InternalServerError"
								},
								"examples": {
									"Example 1": {
										"value": {
											"ok": false,
											"msg": "Error interno del sistema, por favor contacte al administrador del sistema",
											"error": {},
											"status": 500
										}
									}
								}
							}
						}
					}
				},
				"tags": [
					"Tareas"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "tarea_id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateTareaData"
							}
						}
					}
				}
			}
		},
		"/api/tareas/delete/{tarea_id}": {
			"delete": {
				"operationId": "EliminarTarea",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/DeleteTareaResponse"
										},
										{
											"$ref": "#/components/schemas/InternalServerError"
										}
									]
								}
							}
						}
					},
					"500": {
						"description": "Internal Server Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/InternalServerError"
								},
								"examples": {
									"Example 1": {
										"value": {
											"ok": false,
											"msg": "Error interno del sistema, por favor contacte al administrador del sistema",
											"error": {},
											"status": 500
										}
									}
								}
							}
						}
					}
				},
				"tags": [
					"Tareas"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "tarea_id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/api/tareas/mis-tareas": {
			"get": {
				"operationId": "ObtenerMisTareas",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/GetTareasResponse"
										},
										{
											"$ref": "#/components/schemas/InternalServerError"
										}
									]
								}
							}
						}
					},
					"500": {
						"description": "Internal Server Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/InternalServerError"
								},
								"examples": {
									"Example 1": {
										"value": {
											"ok": false,
											"msg": "Error interno del sistema, por favor contacte al administrador del sistema",
											"error": {},
											"status": 500
										}
									}
								}
							}
						}
					}
				},
				"tags": [
					"Tareas"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "empleado_id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "supervisor_id",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
...
		"/api/solicitudes-prestamo/registrar": {
			"post": {
				"operationId": "CreateLoanRequest",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/InternalServerError"
										},
										{
											"$ref": "#/components/schemas/SuccessResponseCreateLoanRequest"
										}
									]
								}
							}
						}
					},
					"500": {
						"description": "Internal Server Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/InternalServerError"
								},
								"examples": {
									"Example 1": {
										"value": {
											"ok": false,
											"msg": "Error interno del sistema, por favor contacte al administrador del sistema",
											"error": {},
											"status": 500
										}
									}
								}
							}
						}
					}
				},
				"tags": [
					"Solicitudes de Préstamo"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/LoanRequest"
							}
						}
					}
				}
			}
		},
		"/api/solicitudes-prestamo/solicitudes/{empresa_id}": {
			"get": {
				"operationId": "GetAllLoanRequestsByCompany",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/InternalServerError"
										},
										{
											"items": {
												"$ref": "#/components/schemas/LoanRequest"
											},
											"type": "array"
										},
										{}
									]
								}
							}
						}
					},
					"500": {
						"description": "Internal Server Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/InternalServerError"
								},
								"examples": {
									"Example 1": {
										"value": {
											"ok": false,
											"msg": "Error interno del sistema, por favor contacte al administrador del sistema",
											"error": {},
											"status": 500
										}
									}
								}
							}
						}
					}
				},
				"tags": [
					"Solicitudes de Préstamo"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "empresa_id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/api/solicitudes-prestamo/solicitudes/{empresa_id}/{id}": {
			"get": {
				"operationId": "GetLoanRequestByIdAndCompany",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/InternalServerError"
										},
										{
											"$ref": "#/components/schemas/LoanRequest"
										}
									]
								}
							}
						}
					},
					"500": {
						"description": "Internal Server Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/InternalServerError"
								},
								"examples": {
									"Example 1": {
										"value": {
											"ok": false,
											"msg": "Error interno del sistema, por favor contacte al administrador del sistema",
											"error": {},
											"status": 500
										}
									}
								}
							}
						}
					}
				},
				"tags": [
					"Solicitudes de Préstamo"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "empresa_id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/api/solicitudes-prestamo/solicitudes/{empresa_id}/{id}/editar": {
			"put": {
				"operationId": "UpdateLoanRequest",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/InternalServerError"
										},
										{
											"$ref": "#/components/schemas/SuccessResponse"
										},
										{}
									]
								}
							}
						}
					},
					"500": {
						"description": "Internal Server Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/InternalServerError"
								},
								"examples": {
									"Example 1": {
										"value": {
											"ok": false,
											"msg": "Error interno del sistema, por favor contacte al administrador del sistema",
											"error": {},
											"status": 500
										}
									}
								}
							}
						}
					}
				},
				"tags": [
					"Solicitudes de Préstamo"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "empresa_id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Partial_LoanRequest_"
							}
						}
					}
				}
			}
		},
...
		"/api/plantillas/plantillas-csv": {
			"post": {
				"operationId": "CrearPlantillaCSV",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/InternalServerError"
										},
										{
											"$ref": "#/components/schemas/SuccessResponse"
										}
									]
								}
							}
						}
					}
				},
				"tags": [
					"Plantillas CSV"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"campos": {
										"items": {
											"properties": {
												"identificador_campo": {
													"type": "string"
												},
												"titulo_campo": {
													"type": "string"
												}
											},
											"required": [
												"identificador_campo",
												"titulo_campo"
											],
											"type": "object"
										},
										"type": "array"
									},
									"titulo_plantilla": {
										"type": "string"
									},
									"empresa_id": {
										"type": "number",
										"format": "double"
									},
									"banco_id": {
										"type": "number",
										"format": "double"
									}
								},
								"required": [
									"campos",
									"titulo_plantilla",
									"empresa_id",
									"banco_id"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/api/plantillas/plantillas-csv/{plantilla_id}": {
			"delete": {
				"operationId": "EliminarPlantillaCSV",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/InternalServerError"
										},
										{
											"$ref": "#/components/schemas/SuccessResponse"
										}
									]
								}
							}
						}
					}
				},
				"tags": [
					"Plantillas CSV"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "plantilla_id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"put": {
				"operationId": "ActualizarPlantillaCSV",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/InternalServerError"
										},
										{
											"$ref": "#/components/schemas/SuccessResponse"
										}
									]
								}
							}
						}
					}
				},
				"tags": [
					"Plantillas CSV"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "plantilla_id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"campos": {
										"items": {
											"properties": {
												"identificador_campo": {
													"type": "string"
												},
												"titulo_campo": {
													"type": "string"
												},
												"campo_id": {
													"type": "number",
													"format": "double"
												}
											},
											"required": [
												"identificador_campo",
												"titulo_campo",
												"campo_id"
											],
											"type": "object"
										},
										"type": "array"
									},
									"titulo_plantilla": {
										"type": "string"
									}
								},
								"required": [
									"campos",
									"titulo_plantilla"
								],
								"type": "object"
							}
						}
					}
				}
			},
			"get": {
				"operationId": "ObtenerPlantillaCSV",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/InternalServerError"
										},
										{
											"$ref": "#/components/schemas/PlantillaCSV"
										},
										{
											"$ref": "#/components/schemas/NotFoundItems"
										}
									]
								}
							}
						}
					}
				},
				"tags": [
					"Plantillas CSV"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "plantilla_id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/api/plantillas/plantillas-csv/empresa/{empresa_id}": {
			"get": {
				"operationId": "ObtenerTodasPlantillasEmpresa",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/InternalServerError"
										},
										{
											"items": {
												"$ref": "#/components/schemas/PlantillaCSV"
											},
											"type": "array"
										}
									]
								}
							}
						}
					}
				},
				"tags": [
					"Plantillas CSV"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "empresa_id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}